Update 1.2 main.py
Changed the stock and price dictionary system into a class system by adding an inventory item class, makes the code much cleaner.

Also separated the buy function into two, one taking an index number and the other taking a name. This makes the process_input() function much cleaner to look at.

Additionally made it so that inputting a quantity of 0 asks you for a quantity.